pyproject.toml
[project]
name = "atomicshabbits"
version = "0.1.0"
description = ""
authors = [
    {name = "Valdemar",email = "vovateslenko.1996.28@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "requests (>=2.32.3,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "eventlet (>=0.39.1,<0.40.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
line_length = 79

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
python_version = "3.11"